mvp integration

+++ Assets/Scripts/UI/CanvasManager.cs
@@ -44,7 +44,7 @@ public class CanvasManager : MonoBehaviour, ICanvasManager
 
     public void UpdateHealth(int cur_health, int max_health)
     {
-        health_slider.GetComponent<Image>().fillAmount = (float)cur_health / max_health;
+        health_slider.GetComponent<Image>().fillAmount = (float)cur_health / (float)max_health;
         health_num.text = cur_health.ToString();
     }
